Move hasCode to the Blockly module to reduce imports.
