+++ src/modules/toolbar_controller.js
@@ -20,9 +20,8 @@
  *   DEALINGS IN THE SOFTWARE.
  */
 
-import Blockly from 'blockly/core';
-
-import {Project, getProjectInitialState} from './project.js';
+import {hasCode} from './blocklyc';
+import {getProjectInitialState} from './project.js';
 import {clientService, serviceConnectionTypes} from './client_service';
 
 
@@ -355,25 +354,4 @@ export function initToolbarIcons() {
     // `Init icons: ${key}, ${value.dataset.icon}, icon:${value.dataset.icon}`);
     $(value).html(bpIcons[value.dataset.icon]);
   });
-}
-
-/**
- * Are there any blocks in the Blockly workspace
- * @return {boolean}
- */
-function hasCode() {
-  let result = false;
-  if (Blockly) {
-    if (Blockly.Xml) {
-      if (Blockly.mainWorkspace) {
-        const xml = Blockly.Xml.workspaceToDom(Blockly.mainWorkspace);
-        const text = Blockly.Xml.domToText(xml);
-        const emptyHeader = Project.getTerminatedEmptyProjectCodeHeader();
-        if (text !== emptyHeader) {
-          result = true;
-        }
-      }
-    }
-  }
-  return result;
 }
